separate file for BatchTable
