+++ src/base/loaders/PNTSLoaderBase.js
@@ -1,7 +1,8 @@
 // PNTS File Format
 // https://github.com/CesiumGS/3d-tiles/blob/master/specification/TileFormats/PointCloud/README.md
 
-import { FeatureTable, BatchTable } from '../../utilities/FeatureTable.js';
+import { BatchTable } from '../../utilities/BatchTable.js';
+import { FeatureTable } from '../../utilities/FeatureTable.js';
 import { readMagicBytes } from '../../utilities/readMagicBytes.js';
 import { LoaderBase } from './LoaderBase.js';
