+++ src/three/loaders/B3DMLoader.d.ts
@@ -1,5 +1,6 @@
 import { B3DMBaseResult } from '../../base/loaders/B3DMLoaderBase';
-import { FeatureTable, BatchTable } from '../../utilities/FeatureTable';
+import { BatchTable } from '../../utilities/BatchTable';
+import { FeatureTable } from '../../utilities/FeatureTable';
 import { LoadingManager, Group } from 'three';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
