+++ src/three/loaders/I3DMLoader.d.ts
@@ -1,5 +1,6 @@
 import { I3DMBaseResult } from '../../base/loaders/I3DMLoaderBase';
-import { FeatureTable, BatchTable } from '../../utilities/FeatureTable';
+import { BatchTable } from '../../utilities/BatchTable';
+import { FeatureTable } from '../../utilities/FeatureTable';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
 import { Group, LoadingManager } from 'three';
