+++ src/utilities/FeatureTable.d.ts
@@ -28,23 +28,3 @@ export class FeatureTable {
 	getBuffer( byteOffset : Number, byteLength : Number ) : ArrayBuffer;
 
 }
-
-export class BatchTable {
-
-	constructor(
-		buffer : ArrayBuffer,
-		batchSize : Number,
-		start : Number,
-		headerLength : Number,
-		binLength : Number
-	);
-
-	getKeys() : Array< String >;
-
-	getData(
-		key : String,
-		componentType : String | null,
-		type : String | null
-	) : Number | String | ArrayBufferView;
-
-}
